Display user, edit profile modal, delete account alert

+++ src/screens/ProfileScreen.jsx
@@ -1,5 +1,5 @@
-import React, { useContext } from "react";
-import { View, StyleSheet } from "react-native";
+import React, { useEffect, useContext } from "react";
+import { View, StyleSheet, Alert } from "react-native";
 import { theme } from "../theme";
 
 import { AuthenticationContext } from "../context/AuthenticationContext";
@@ -8,29 +8,56 @@ import { UserContext } from "../context/UserContext";
 import CustomButton from "./../components/CustomButton";
 import UserProfileCard from "../components/UserProfileCard";
 //TODO: fetch users information
-const ProfileScreen = () => {
+const ProfileScreen = ({navigation}) => {
   const { user, onLogout } = useContext(AuthenticationContext);
-  const { onDeleteUser } = useContext(UserContext);
+  const { getUser, userInfo, onDeleteUser } = useContext(UserContext);
 
-  const deleteHandler = () => {
-    onDeleteUser(user.uid);
-    onLogout();
-  };
+  useEffect(() => {
+    getUser(user.uid);
+  }, []);
+
+  const onEditProfileInfo = () => navigation.navigate("EditProfileInfo");
+
+  const deleteHandler = () =>
+    Alert.alert(
+      "Delete Account!",
+      "Are you sure you want to delete your account",
+      [
+        {
+          text: "Cancel",
+          style: "cancel",
+        },
+        {
+          text: "Delete",
+          onPress: () => {
+            onDeleteUser(user.uid);
+            onLogout();
+          },
+        },
+      ]
+    );
+
   return (
     <View style={styles.container}>
       <View style={styles.cardContainer}>
         <UserProfileCard
-          firstName="Kurosh"
-          lastName="Husseini"
-          email="admin@gmail.com"
-          number="045682266676"
+          firstName={userInfo?.firstName}
+          lastName={userInfo?.lastName}
+          email={userInfo?.email}
+          number={userInfo?.phoneNumber}
         />
       </View>
       <View style={styles.buttonContainer}>
         <View style={styles.innerBtnContainer}>
           <CustomButton
             title="edit your information"
-            onPressHandler={() => onLogout()}
+            onPressHandler={onEditProfileInfo}
+          />
+        </View>
+        <View style={styles.innerBtnContainer}>
+          <CustomButton
+            title="change your password"
+            onPressHandler={() => console.log("change password")}
           />
         </View>
         <View style={styles.innerBtnContainer}>
